Refactor homepage into modular sections
Split the main page into modular components: HeroSection, AboutSection, ServicesSection, ProjectsPreviewSection, and ContactSection. Updated imports and props passing in page.js for improved maintainability. Removed 'Services' from Header navigation and adjusted section references. This refactor simplifies the homepage structure and improves code organization.
